Adjust EntitySerializeTrait to make both Serializable & JsonSerializable implementable
