+++ src/Storage/Entity/Entity.php
@@ -17,6 +17,9 @@ abstract class Entity implements ArrayAccess, JsonSerializable
 
     /** @var int */
     protected $id;
+    /** @var array */
+    protected $_internal = ['contenttype'];
+
 
     /**
      * Constructor.
@@ -73,8 +76,36 @@ abstract class Entity implements ArrayAccess, JsonSerializable
         $this->id = $id;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function __toString()
     {
         return (string) $this->getId();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function toArray()
+    {
+        $data = [];
+        foreach ($this as $k => $v) {
+            if (strpos($k, '_') === 0) {
+                continue;
+            }
+            if (in_array($k, $this->_internal)) {
+                continue;
+            }
+            $method = 'serialize' . $k;
+            $data[$k] = $this->$method();
+        }
+
+        foreach ($this->_fields as $k => $v) {
+            $method = 'serialize' . $k;
+            $data[$k] = $this->$method();
+        }
+
+        return $data;
     }
 }
